backend/requirements.txt
fastapi==0.115.9
uvicorn==0.23.2
sqlalchemy==2.0.22
alembic==1.12.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pymupdf==1.23.8
langchain-core==0.1.52
langchain==0.1.20
langchain-openai==0.1.8
langchain-community==0.0.38
langchain-chroma==0.1.2
langchain-text-splitters==0.0.1
chromadb==0.4.24
openai==1.78.0
redis==5.0.1
celery==5.3.4
stripe==7.3.0
psycopg2-binary==2.9.9
structlog==23.2.0
sentry-sdk==1.32.0
slowapi==0.1.8
python-dotenv==1.0.0
PyJWT==2.8.0
asyncpg==0.30.0
weaviate-client==3.25.1
google-generativeai==0.3.2 